+++ core/src/main/java/org/jdbi/v3/core/mapper/reflect/ConstructorInstanceFactory.java
@@ -14,7 +14,6 @@
 package org.jdbi.v3.core.mapper.reflect;
 
 import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
@@ -24,28 +23,31 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.WeakHashMap;
+import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Stream;
 
-import org.jdbi.v3.core.internal.exceptions.Sneaky;
+import org.jdbi.v3.core.config.ConfigRegistry;
+import org.jdbi.v3.core.config.internal.ConfigCache;
+import org.jdbi.v3.core.config.internal.ConfigCaches;
+import org.jdbi.v3.core.internal.JdbiClassUtils;
+import org.jdbi.v3.core.internal.exceptions.Unchecked;
 
 import static java.util.Collections.synchronizedMap;
 import static java.util.Objects.requireNonNull;
 
 class ConstructorInstanceFactory<T> extends InstanceFactory<T> {
-    private static final MethodHandles.Lookup LOOKUP = MethodHandles.lookup();
-    private static final Map<Constructor<?>, ConstructorHandleAndTypes> CONSTRUCTOR_CACHE = synchronizedMap(new WeakHashMap<>());
+    private static final Map<Constructor<?>, List<Type>> TYPES_CACHE = synchronizedMap(new WeakHashMap<>());
+    private static final ConfigCache<Constructor<?>, MethodHandle> HANDLE_CACHE =
+            ConfigCaches.declare((config, ctor) -> JdbiClassUtils.unreflectConstructor(config, ctor).asFixedArity());
 
     private final Constructor<T> constructor;
     private final List<Type> types;
-    private final MethodHandle constructorHandle;
 
     ConstructorInstanceFactory(Constructor<T> constructor) {
         super(constructor);
         this.constructor = requireNonNull(constructor, "constructor is null");
-        ConstructorHandleAndTypes constructorHandleAndTypes = getConstructorHandleAndTypes(constructor, super::getTypes);
-        this.types = constructorHandleAndTypes.getTypes();
-        this.constructorHandle = constructorHandleAndTypes.getConstructorHandle();
+        this.types = TYPES_CACHE.computeIfAbsent(constructor, ctor -> computeTypes(ctor, super::getTypes));
     }
 
     @Override
@@ -55,12 +57,9 @@ class ConstructorInstanceFactory<T> extends InstanceFactory<T> {
 
     @SuppressWarnings("unchecked")
     @Override
-    T newInstance(Object... params) {
-        try {
-            return (T) constructorHandle.invokeWithArguments(params);
-        } catch (Throwable e) {
-            throw Sneaky.throwAnyway(e);
-        }
+    Function<Object[], T> instantiator(ConfigRegistry config) {
+        final MethodHandle handle = HANDLE_CACHE.get(constructor, config);
+        return Unchecked.function(params -> (T) handle.invokeWithArguments(params));
     }
 
     @Override
@@ -92,48 +91,17 @@ class ConstructorInstanceFactory<T> extends InstanceFactory<T> {
         return lossDetected;
     }
 
-    private static <T> ConstructorHandleAndTypes getConstructorHandleAndTypes(Constructor<T> constructor, Supplier<List<Type>> defaultSupplier) {
-        return CONSTRUCTOR_CACHE.computeIfAbsent(constructor, ctor -> computeConstructorHandleAndTypes(ctor, defaultSupplier));
-    }
-
-    private static <T> ConstructorHandleAndTypes computeConstructorHandleAndTypes(Constructor<T> constructor, Supplier<List<Type>> defaultSupplier) {
-        MethodHandle constructorMethodHandle = getConstructorMethodHandle(constructor);
+    private static List<Type> computeTypes(Constructor<?> constructor, Supplier<List<Type>> defaultSupplier) {
         if (isGenericInformationLost(constructor)) {
-            return new ConstructorHandleAndTypes(constructorMethodHandle, getFields(constructor)
+            return getFields(constructor)
                 .map(Field::getGenericType)
-                .toList());
+                .toList();
         }
-        return new ConstructorHandleAndTypes(constructorMethodHandle, defaultSupplier.get());
-    }
-
-    private static <T> MethodHandle getConstructorMethodHandle(Constructor<T> constructor) {
-        try {
-            return LOOKUP.unreflectConstructor(constructor).asFixedArity();
-        } catch (IllegalAccessException e) {
-            throw new RuntimeException(e);
-        }
+        return defaultSupplier.get();
     }
 
     private static <T> Stream<Field> getFields(Constructor<T> constructor) {
         return Arrays.stream(constructor.getDeclaringClass().getDeclaredFields())
             .filter(field -> !Modifier.isStatic(field.getModifiers()));
-    }
-
-    private static class ConstructorHandleAndTypes {
-        private final MethodHandle constructorHandle;
-        private final List<Type> types;
-
-        ConstructorHandleAndTypes(MethodHandle constructorHandle, List<Type> types) {
-            this.constructorHandle = requireNonNull(constructorHandle, "constructorHandle is null");
-            this.types = requireNonNull(types, "types is null");
-        }
-
-        public MethodHandle getConstructorHandle() {
-            return constructorHandle;
-        }
-
-        public List<Type> getTypes() {
-            return types;
-        }
     }
 }